Msg ao deletar "Categoria já foi removida" para ID que já foi removido.

+++ src/main/java/jdev/mentoria/lojavirtual/controller/CategoriaProdutoControle.java
@@ -37,6 +37,10 @@ public class CategoriaProdutoControle {
 	@PostMapping(value = "**/deleteCategoria") /*Mapeando a url para receber JSON*/
 	public ResponseEntity<?> deleteAcesso(@RequestBody CategoriaProduto categoriaProduto) { /*Recebe o JSON e converte pra Objeto*/
 		
+		if (categoriaProdutoRepository.findById(categoriaProduto.getId()).isPresent() == false) {
+			return new ResponseEntity("Categoria já foi removida",HttpStatus.OK);
+		}
+		
 		categoriaProdutoRepository.deleteById(categoriaProduto.getId());
 		
 		return new ResponseEntity("Categoria Removida",HttpStatus.OK);
